+++ pytsite/auth/driver/abstract.py
@@ -10,7 +10,7 @@ __license__ = 'MIT'
 
 class AbstractDriver(_ABC):
     @_abstractmethod
-    def get_login_form(self, uid: str=None, css: str=None, legend: str=None) -> _form.Base:
+    def get_login_form(self, uid='', css='', title='') -> _form.Base:
         """Login form get handler.
         """
         raise NotImplementedError()