+++ pytsite/content/_form.py
@@ -11,31 +11,32 @@ class Settings(_form.Base):
     """Content Settings Form.
     """
     def _setup(self):
-        i = 1
+        i = 10
         for lang_code in _lang.get_langs():
             self.add_widget(_widget.input.Text(
                 uid='setting_home_title_' + lang_code,
                 label=_lang.t('pytsite.content@home_page_title', language=lang_code),
-                weight=10 * i,
+                weight=i,
             ))
+            i += 10
 
             self.add_widget(_widget.input.Text(
                 uid='setting_home_description_' + lang_code,
                 label=_lang.t('pytsite.content@home_page_description', language=lang_code),
-                weight=20 * i,
+                weight=i,
             ))
+            i += 10
 
             self.add_widget(_widget.input.Tokens(
                 uid='setting_home_keywords_' + lang_code,
                 label=_lang.t('pytsite.content@home_page_keywords', language=lang_code),
-                weight=30 * i,
+                weight=i,
             ))
-
-            i += 1
+            i += 10
 
         self.add_widget(_widget.input.TextArea(
             uid='setting_add_js',
             label=_lang.t('pytsite.content@additional_js_code'),
             rows=10,
-            weight=100,
+            weight=i,
         ))