+++ pytsite/html/__init__.py
@@ -394,7 +394,7 @@ class TextArea(BlockElement):
 
 class Label(InlineElement):
     def _get_valid_attrs(self) -> tuple:
-        return 'label_for',
+        return 'label_for', 'class'
 
 
 class Select(BlockElement):