+++ pytsite/settings/_model.py
@@ -1,10 +1,10 @@
 """Settings Models.
 """
+from pytsite import odm as _odm
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
-from pytsite import odm as _odm
 
 
 class Setting(_odm.Model):